Update more packets

+++ src/main/java/cn/nukkit/network/protocol/PlayerHotbarPacket.java
@@ -7,8 +7,7 @@ import lombok.ToString;
 public class PlayerHotbarPacket extends DataPacket {
 
     public int selectedHotbarSlot;
-    public int windowId = ContainerIds.INVENTORY;
-
+    public byte windowId = ContainerIds.INVENTORY;
     public boolean selectHotbarSlot = true;
 
     @Override
@@ -18,7 +17,7 @@ public class PlayerHotbarPacket extends DataPacket {
 
     @Override
     public void decode() {
-        this.selectedHotbarSlot = (int) this.getUnsignedVarInt();
+        this.selectedHotbarSlot = this.getUnsignedVarInt();
         this.windowId = this.getByte();
         this.selectHotbarSlot = this.getBoolean();
     }
@@ -27,7 +26,7 @@ public class PlayerHotbarPacket extends DataPacket {
     public void encode() {
         this.reset();
         this.putUnsignedVarInt(this.selectedHotbarSlot);
-        this.putByte((byte) this.windowId);
+        this.putByte(this.windowId);
         this.putBoolean(this.selectHotbarSlot);
     }
 }
